requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.0
PyMySQL==1.1.0
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
plotly==5.18.0

